requirements_live.txt
ccxt>=4.0
python-dotenv>=1.0
pandas>=2.0
numpy>=1.24
TA-Lib>=0.4.28
requests>=2.31
